Update news-management.en.md
I updated this one. I don't know how to updated the images. I'm not super familiar with Github. Can you give me some direction for that, or if it's easier feel free. It would be good for me to learn though. I have been going in once and while and submitting updates in other spots of the documentation.